AB#1271 move GraphQLService to ws folder, test expired token
