+++ src/main/java/mil/dds/anet/utils/DaoUtils.java
@@ -15,8 +15,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
-import mil.dds.anet.beans.Assessment;
 import mil.dds.anet.beans.AccessToken;
+import mil.dds.anet.beans.Assessment;
 import mil.dds.anet.beans.CustomSensitiveInformation;
 import mil.dds.anet.beans.Person;
 import mil.dds.anet.beans.Position;
